+++ src/test/java/com/back/domain/member/controller/MemberControllerTest.java
@@ -1,4 +1,4 @@
-package com.back.backend.domain.member.controller;
+package com.back.domain.member.controller;
 
 import com.back.domain.member.member.entity.Member;
 import com.back.domain.member.member.repository.MemberRepository;
