+++ src/test/java/com/back/domain/quiz/daily/controller/DailyQuizControllerTest.java
@@ -1,7 +1,7 @@
-package com.back.backend.domain.quiz.daily.controller;
+package com.back.domain.quiz.daily.controller;
 
-import com.back.backend.global.config.TestRqConfig;
-import com.back.backend.global.rq.TestRq;
+import com.back.global.config.TestRqConfig;
+import com.back.global.rq.TestRq;
 import com.back.domain.member.member.entity.Member;
 import com.back.domain.member.member.service.MemberService;
 import com.back.global.rq.Rq;
