+++ src/test/java/com/back/domain/quiz/detail/controller/DetailQuizControllerTest.java
@@ -1,7 +1,7 @@
-package com.back.backend.domain.quiz.detail.controller;
+package com.back.domain.quiz.detail.controller;
 
-import com.back.backend.global.config.TestRqConfig;
-import com.back.backend.global.rq.TestRq;
+import com.back.global.config.TestRqConfig;
+import com.back.global.rq.TestRq;
 import com.back.domain.member.member.entity.Member;
 import com.back.domain.member.member.service.MemberService;
 import com.back.domain.quiz.detail.dto.DetailQuizDto;
